Change dynamic dependency of FileUpload in Messages models (#13776)
* Change dynamic dependency of FileUpload in Messages models

* Move require to top

* Fix lint

* Move file upload from lib to server only

* Fix wrong import
