+++ app/file-upload/client/lib/fileUploadHandler.js
@@ -4,22 +4,15 @@ import { Tracker } from 'meteor/tracker';
 import { UploadFS } from 'meteor/jalik:ufs';
 import { FileUploadBase } from '../../lib/FileUploadBase';
 import { Uploads, Avatars } from '../../../models';
-import { FileUpload } from '../../lib/FileUpload';
 
 new UploadFS.Store({
 	collection: Uploads.model,
 	name: 'Uploads',
-	filter: new UploadFS.Filter({
-		onCheck: FileUpload.validateFileUpload,
-	}),
 });
 
 new UploadFS.Store({
 	collection: Avatars.model,
 	name: 'Avatars',
-	filter: new UploadFS.Filter({
-		onCheck: FileUpload.validateFileUpload,
-	}),
 });
 
 export const fileUploadHandler = (directive, meta, file) => {
