+++ app/settings/server/functions/settings.js
@@ -1,6 +1,6 @@
 import { Meteor } from 'meteor/meteor';
 import { settings } from '../../lib/settings';
-import { Settings } from '../../../models';
+import Settings from '../../../models/server/models/Settings';
 import _ from 'underscore';
 
 const blockedSettings = {};
